requirements.txt
google-cloud-bigquery>=3.30.0
google-auth>=2.48.0
gspread>=6.2.1
